Implement Set User Role <Domain><Infras-Migration><Application><Facade><Api>

+++ Shop/Shop.Domain/UserAgg/UserRole.cs
@@ -4,12 +4,12 @@ namespace Shop.Domain.UserAgg
 {
     public class UserRole : BaseEntity
     {
-        public UserRole(int roleId)
+        public UserRole(long roleId)
         {
             RoleId = roleId;
         }
 
         public long UserId { get; internal set; }
-        public int RoleId { get; private set; }
+        public long RoleId { get; private set; }
     }
 }
